+++ src/BackupJob.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Spatie\Backup;
+
+class BackupJob
+{
+    public function __construct()
+    {
+    }
+
+    public static function create() : BackupJob
+    {
+        return new static();
+    }
+
+    public function doNotIncludeAnyFiles()
+    {
+        return $this;
+    }
+
+    public function run()
+    {
+    }
+}