+++ tests/Integration/TestCase.php
@@ -1,0 +1,115 @@
+<?php
+
+namespace Spatie\Backup\Test\Integration;
+
+use File;
+use Illuminate\Database\Schema\Blueprint;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+abstract class TestCase extends Orchestra
+{
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->setUpDatabase($this->app);
+
+        $this->setUpTempTestFiles();
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app)
+    {
+        return [
+            \Spatie\Backup\BackupServiceProvider::class,
+        ];
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     */
+    protected function getEnvironmentSetUp($app)
+    {
+        $this->initializeDirectory($this->getTempDirectory());
+
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite', [
+            'driver' => 'sqlite',
+            'database' => $this->getTempDirectory().'/database.sqlite',
+            'prefix' => '',
+        ]);
+
+        $app['config']->set('filesystems.disks.media', [
+            'driver' => 'local',
+            'root' => $this->getMediaDirectory(),
+        ]);
+
+        $app['config']->set('filesystems.disks.secondMediaDisk', [
+            'driver' => 'local',
+            'root' => $this->getTempDirectory('media2'),
+        ]);
+
+        $app->bind('path.public', function () {
+            return $this->getTempDirectory();
+        });
+
+        $app['config']->set('app.key', '6rE9Nz59bGRbeMATftriyQjrpF7DcOQm');
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     */
+    protected function setUpDatabase($app)
+    {
+        file_put_contents($this->getTempDirectory().'/database.sqlite', null);
+
+        $app['db']->connection()->getSchemaBuilder()->create('test_models', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('name');
+        });
+
+        TestModel::create(['name' => 'test']);
+
+        include_once '__DIR__'.'/../resources/migrations/create_media_table.php.stub';
+
+        (new \CreateMediaTable())->up();
+    }
+
+    protected function setUpTempTestFiles()
+    {
+        $this->initializeDirectory($this->getTestFilesDirectory());
+        File::copyDirectory(__DIR__.'/../resources/testfiles', $this->getTestFilesDirectory());
+    }
+
+    protected function initializeDirectory($directory)
+    {
+        if (File::isDirectory($directory)) {
+            File::deleteDirectory($directory);
+        }
+        File::makeDirectory($directory);
+    }
+
+    public function getTempDirectory($suffix = '')
+    {
+        return __DIR__.'/temp'.($suffix == '' ? '' : '/'.$suffix);
+    }
+
+    public function getMediaDirectory($suffix = '')
+    {
+        return $this->getTempDirectory().'/media'.($suffix == '' ? '' : '/'.$suffix);
+    }
+
+    public function getTestFilesDirectory($suffix = '')
+    {
+        return $this->getTempDirectory().'/testfiles'.($suffix == '' ? '' : '/'.$suffix);
+    }
+
+    public function getTestJpg()
+    {
+        return $this->getTestFilesDirectory('test.jpg');
+    }
+}