+++ tests/Unit/ExampleTest.php
@@ -1,13 +1,11 @@
 <?php
 
-namespace Spatie\Skeleton\Test;
+namespace Spatie\Skeleton\Test\Unit;
 
 class ExampleTest extends \PHPUnit_Framework_TestCase
 {
-
-
     /**
-     * Test that true does in fact equal true
+     * Test that true does in fact equal true.
      */
     public function testTrueIsTrue()
     {